Part 2A: add teachers scaffold
This adds a Teacher model and associated routes for basic CRUD.  The
teacher is very simple and only has a name field.  In order to keep
things DRY, Student name validation is moved into a Person module which
is included in both models.

There's also a bunch of refactoring of forms and specs to keep those
somewhat DRY also.